+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.0] - 2020-12-07
+
+### Added
+
 - Default the Operator Version Label in `Cluster` to match the new release during upgrade.
 - Default the Release Version Label in `Cluster` to the newest active production release.
 
@@ -163,7 +167,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/aws-admission-controller/compare/v2.5.0...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-admission-controller/compare/v2.6.0...HEAD
+[2.6.0]: https://github.com/giantswarm/aws-admission-controller/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/giantswarm/aws-admission-controller/compare/v2.4.1...v2.5.0
 [2.4.1]: https://github.com/giantswarm/aws-admission-controller/compare/v2.4.0...v2.4.1
 [2.4.0]: https://github.com/giantswarm/aws-admission-controller/compare/v2.3.3...v2.4.0